requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-filter==2.4.0
djangorestframework==3.12.2
idna==2.10
importlib-metadata==3.4.0
install==1.3.3
Js2Py==0.70
Markdown==3.3.3
numpy==1.19.0
pandas==1.0.5
pyjsparser==2.7.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
twilio==6.44.0
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.9
zipp==3.4.0
